refactor (gravsearch): move local methods from SearchResponderV2 to GravsearchUtilV2
